Fixing sigma reads

+++ src/dark_matter_io.h
@@ -187,7 +187,7 @@ INLINE static int sidm_write_dmparts(const struct dmpart* dmparts,
     list[6] = io_make_output_field("Max_SIDM_events_per_timestep", FLOAT, 1, UNIT_CONV_NO_UNITS, 0.f,
                                    dmparts, sidm_data.max_sidm_events_per_timestep, "Maximum number of DM-DM collisions the particle has had in a single timestep");
 
-    list[7] = io_make_output_field("Cross_section", DOUBLE, 1, UNIT_CONV_NO_UNITS, 0.f, dmparts, sidm_data.sigma,"DM particle cross section [internal units]");
+    list[7] = io_make_output_field("Cross_section", FLOAT, 1, UNIT_CONV_NO_UNITS, 0.f, dmparts, sidm_data.sigma,"DM particle cross section [internal units]");
                                    
     return 8;
